move message-service-related files to extensions
