+++ src/extensions/message-handlers/initaltext-guard/index.js
@@ -1,6 +1,4 @@
-import { getConfig } from "../../../server/api/lib/config";
-import { r, cacheableData, Message } from "../../../server/models";
-import serviceMap from "../../../server/api/lib/services";
+import { r } from "../../../server/models";
 
 export const serverAdministratorInstructions = () => {
   return {
