+++ src/server/index.js
@@ -3,7 +3,7 @@ import bodyParser from "body-parser";
 import express from "express";
 import appRenderer from "./middleware/app-renderer";
 import { graphqlExpress, graphiqlExpress } from "apollo-server-express";
-import { makeExecutableSchema, addMockFunctionsToSchema } from "graphql-tools";
+import { makeExecutableSchema } from "graphql-tools";
 // ORDERING: ./models import must be imported above ./api to help circular imports
 import { createLoaders, createTablesIfNecessary, r } from "./models";
 import { resolvers } from "./api/schema";
@@ -14,8 +14,8 @@ import passportSetup from "./auth-passport";
 import wrap from "./wrap";
 import { log } from "../lib";
 import telemetry from "./telemetry";
-import nexmo from "./api/lib/nexmo";
-import twilio from "./api/lib/twilio";
+import nexmo from "../extensions/messaging_services/nexmo";
+import twilio from "../extensions/messaging_services/twilio";
 import { getConfig } from "./api/lib/config";
 import { seedZipCodes } from "./seeds/seed-zip-codes";
 import { setupUserNotificationObservers } from "./notifications";
